Move Duration to libcore.
