+++ src/libcoretest/lib.rs
@@ -71,4 +71,5 @@ mod ptr;
 mod result;
 mod slice;
 mod str;
+mod time;
 mod tuple;
